+++ tests/lib/utils/selector.test.ts
@@ -13,7 +13,7 @@ function loadPatterns() {
       path.join(FIXTURE_ROOT, name, 'source.vue'),
       'utf8'
     )
-    const code = code0.replace(/^<!--(.+?)-->/, `<!--${name}-->`)
+    const code = code0.replace(/^<!--(.+?)-->/, () => `<!--${name}-->`)
     const inputSelector = /^<!--(.+?)-->/.exec(code0)![1].trim()
     return { code, name, inputSelector }
   })
@@ -61,11 +61,11 @@ function extractElements(code: string, inputSelector: string) {
 describe.each(loadPatterns())(
   `parseSelector() [$name]`,
   ({ name, code, inputSelector }) => {
-    it('should to parse the selector to match the valid elements.', () => {
+    it('should to parse the selector to match the valid elements.', async () => {
       const elements = extractElements(code, inputSelector)
       const actual = JSON.stringify(elements, null, 4)
 
-      expect(actual).toMatchFileSnapshot(
+      await expect(actual).toMatchFileSnapshot(
         path.join(FIXTURE_ROOT, name, 'result.json')
       )
     })